+++ sys/build.rs
@@ -15,10 +15,10 @@ use syn::visit::Visit;
 /// Version of the used [TigerBeetle] release.
 ///
 /// [TigerBeetle]: https://github.com/tigerbeetle/tigerbeetle
-const TIGERBEETLE_RELEASE: &str = "0.16.11";
+const TIGERBEETLE_RELEASE: &str = "0.16.17";
 
 /// Commit hash of the [`TIGERBEETLE_RELEASE`].
-const TIGERBEETLE_COMMIT: &str = "ea8a3e445fd1801d8f5ad1dbd6a9320861053912";
+const TIGERBEETLE_COMMIT: &str = "19a37355a64d09b0e35b14e5d1699e098bffdab8";
 
 fn target_to_lib_dir(target: &str) -> Option<&'static str> {
     match target {
@@ -54,11 +54,14 @@ const SCRIPT_EXTENSION: &str = "bat";
 fn main() {
     assert!(env!("CARGO_PKG_VERSION").ends_with(TIGERBEETLE_RELEASE));
     let out_dir: PathBuf = env::var("OUT_DIR").unwrap().into();
-    let debug: bool = env::var("DEBUG").unwrap().parse().unwrap();
+    let debug: bool = env::var("TB_CLIENT_DEBUG").map_or_else(
+        |_| env::var("DEBUG").unwrap().parse().unwrap(),
+        |s| s.parse().unwrap(),
+    );
     let target = env::var("TARGET").unwrap();
-    let log_level = env::var("TIGERBEETLE_LOG_LEVEL").unwrap_or_else(|_| "info".to_owned());
 
     println!("cargo:rerun-if-env-changed=DOCS_RS");
+    println!("cargo:rerun-if-env-changed=TB_CLIENT_DEBUG");
     println!("cargo:rerun-if-changed=src/wrapper.h");
 
     let wrapper;
@@ -110,7 +113,6 @@ fn main() {
         .arg("clients:c")
         .args((!debug).then_some("-Drelease"))
         .arg(format!("-Dtarget={tigerbeetle_target}"))
-        .arg(format!("-Dconfig-log-level={log_level}"))
         .arg(format!("-Dconfig-release={TIGERBEETLE_RELEASE}"))
         .arg(format!("-Dconfig-release-client-min={TIGERBEETLE_RELEASE}"))
         .arg(format!("-Dgit-commit={TIGERBEETLE_COMMIT}"))